move hourly breakdown report generator to helpers
