+++ src/helpers/generate-hourly-breakdown-ephemeral-report/index.ts
@@ -1,0 +1,18 @@
+export function generateHourlyBreakdownEphemeralReport(space, startDate, endDate) {
+  return {
+    id: `rpt_${space.id}`,
+    type: 'HOURLY_BREAKDOWN',
+    name: 'Hourly Breakdown',
+    settings: {
+      "spaceId": space.id,
+      "timeRange": {
+        type: 'CUSTOM_RANGE',
+        startDate,
+        endDate,
+      },
+      "includeWeekends": true,
+      "hourStart": 6,
+      "hourEnd": 20
+    },
+  };
+}
